reinteract_wrapper_osx: use unsigned rather than NSUInteger
NSUInteger seems to be newer than the 10.4 sdk, so replace it
with a simple 'unsigned'.

+++ src/reinteract_wrapper_osx/main.m
@@ -77,7 +77,7 @@ fileExists(NSString *path)
 static PyObject *
 string_to_unicode(NSString *string)
 {
-    NSUInteger length = [string length];
+    unsigned length = [string length];
     Py_UNICODE *buffer = malloc(length * sizeof(Py_UNICODE));
     PyObject *result;
     if (!buffer)
